Add the exporter for client
